+++ lna_photo/lib/Views/image_detail_view/image_detail_view.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+import 'package:lna_photo/Models/photo.dart';
+
+import 'image_detail.dart';
+
+class ImageDetailView extends StatefulWidget {
+  const ImageDetailView({Key? key, required this.photo}) : super(key: key);
+
+  final Photo photo;
+
+  @override
+  State<ImageDetailView> createState() => _ImageDetailViewState();
+}
+
+class _ImageDetailViewState extends State<ImageDetailView> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: const Color.fromARGB(255, 1, 93, 105),
+        title: Text(widget.photo.name),
+      ),
+      body: ImageDetail(
+        photo: widget.photo,
+      ),
+    );
+  }
+}