+++ .github/Changelog.md
@@ -1,1 +1,5 @@
 # Changelog 🪵
+
+- Added loading indicator to highres images
+- Added highres images to detail view
+- Added an image detail view